feat: add solution task module
